+++ packages/firebase-auth/src/cloudflare-types.d.ts
@@ -1,0 +1,5 @@
+declare namespace Cloudflare {
+  interface Env {
+    PUBLIC_JWK_CACHE_KV: KVNamespace
+  }
+}